Fix SafeDestroy when entering a numeric resource ID
In a previous commit, the inputField of SafeDestroy.js was converted
from a numberfield to a textfield to allow ceph pools names to be
entered.
This means the type of 'value' changed from number to string and
the strong equal test === was failing against a numeric id.
To solve this, we now always stringify the resource id before comparing.