fix wizard validity check for options spanning multiple tabs
when changing the guest os, we changed the disk type
(e.g. ide for windows, scsi for linux, etc.)

but if the id was outside the allowed range (e.g. > 3 for ide)
we did not correctly enable/disable tabs

now we check all tabs until the highest already visited,
or until we detect a not valid tab

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Thomas Lamprecht <t.lamprecht@proxmox.com>