postinst: remove fixup for ceph jewel update transition
In commit 8b89f7533e2df983e0285e2f170971a3637d7349 we needed to
change the ceph.service install target to multi-user.target, as
ceph.target could hang indefinitely if ceph-common gets upgraded.
This change is included in pve-manager 4.4-13 and newer, as users
wanting to upgrade to 5.0 must upgrade to latest 4.4 to be able to do
so (without headache) this can be removed.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>