ui/acme: remove 'Renew Cert' button for now
You can normally do this with the Order Cert button, with almost the
same result, and it isn't something which will be needed often, if
the UI is somehow inaccessible (due to cert problem) then this is
useless anyway, and the CLI must be used anyhow.

Can be easily reverted if we deem this necessary later on.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>