ui: dc/storage: allow to renenable iscsi storages again
The dc/storage refactoring done in
commit 8f60ee4ccacc14f1bbe025874522c151cc375f83
introduced a regression where an once deactivated iSCSI storage could
not be activated again through UI, as we did not called the parent
setValue function correctly, thus the Enable checkbox was always set
to true. A uncheck -> check cycle couldn't also be done as the dirty
changes check disabled the submit button in this case.

Enable/Disable through API or CLI still worked.

Reported-by: Bastian Sebode <b.sebode@linet-services.de>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>