ui/storage: cifs: reoder fields
username, password and domain name are all related, they are non-pve
options related to accessing the CIFS storage and have all influence
on the pveCIFSScan component.
So group at kleast user and password together. But keep domain on the
right for the sake of balance.
Move the content type selector to the right, as the RBD and other
panels do.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>