ui: storage: combine RBD external and hyperconverged add dialog
Combine both dialogues. This reduces storage choices from the
Storage -> Add menu, and thus improves usability.

We make the whole dialog also more intelligent by querying the
monitors, so that we can show them for the hyper converged case
keeping a similar interface for both cases. Further we can use this
to decide if a hyperconverged setup is possible at all.
While this information would also be available through the fact if
the pool selector has entries I wanted to show them anyway and could
make logic a bit easier using it.