5to6: get_pkg: cache version list
this is a short running script, so the version list can be re-used,
the chance that there where updates in between are slim and racy
anyway. IF getting the versions did not succeeded, we still retry on
every call though, simpler and ensures a warning is printed in the
caller check vicinity.

Makes script noticeable faster.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>