fix #1594: add "Run now" button to cluster backup page
Iterate all (online) nodes client-side and call vzdump with the correct
parameters (according to the job selected) for each one.

Then, show a progress bar in a non-closeable modal-dialog, to ensure the
user stays on the Backup page during the /vzdump API calls. Any errors
that occurred will be displayed in a consolidated message box.

Includes a "confirm" dialog to not accidentally run a potentially large
backup job.

Curiously, the "pve-cluster-backup" data model seems to have been broken
entirely.  I'm not sure how it was working before, but the changes in
this patch need it fixed, so include that as well (use "mode" instead of
individual flags + "compress" is not a boolean).

Signed-off-by: Stefan Reiter <s.reiter@proxmox.com>