ui: backup restore: add start-after-restore checkbox
Restore may be a long running operation, as it goes over the common
create API path we can reuse the "start after create" for a "start
after restore" for free.

Add this as hbox layout, as else it looks like a lot of wasted space
in the right side of the window, especially when restoring a CT
backup (were the "unprivileged" checkbox is additionally there).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>